refactor: Modify exception handling to be delegated to Service
